% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglfm.R
\name{biglfm}
\alias{biglfm}
\title{Joint maximum likelihood estimator for general latent factor model fitting to large scale dataset.}
\usage{
biglfm(model, response, theta0, A0, Q, C2, tol = 1e-04, parallel = T)
}
\arguments{
\item{model}{one of the three models: "linear", "mirt", "poisson" must be selected.}

\item{response}{N by J 0-1 matrix, each row represents test takers' response to J items.}

\item{theta0}{N by K matrix, initial values of latent traits parameters.}

\item{A0}{J by K matrix, initial values of loading matrix paramters.}

\item{Q}{J by K matrix, pre-specified Q matrix for structured information, if Q is set to NULL, the algorithm performs 
exploretory setting.}

\item{C2}{Constraint constant, for more information see to the paper \url{www.google.com}.}

\item{tol}{Tolerance threshold, default value is 1e-4.}

\item{parallel}{If enable parallel computing, default value is True, i.e. use all the computing cores in your computer.}
}
\value{
A list contains the estimated loading matrix A and latent traits matrix theta.
}
\description{
Joint maximum likelihood estimator for general latent factor model fitting to large scale dataset.
}
\examples{
\dontrun{
library(biglfm)
# To fit the Linear model
res_linear <- biglfm(model = "linear", response_linear,
 theta, A, Q, 1.44*C2, tol = 1e-4, parallel = T)

# MIRT model
res_mirt <- biglfm(model = "mirt",response_mirt, theta, A, Q, 1.44*C2, tol = 1e-4, parallel =T)

# Poisson modeel
res_poi <- biglfm(model = "poisson",response_poi, theta, A, Q, 1.44*C2, tol = 1e-4, parallel = T)
}
}
