% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Responses of 2500 people for 100 items with 5 latent traits.}
\format{A list with 7 elements:
\describe{
   \item{A}{loading matrix, 100 by 5.}
   \item{theta}{latent traits matrix, 2500 by 5.}
   \item{Q}{Q matrix, 100 by 5 with elements 0 or 1.}
   \item{C2}{squared constraint constant, double.}
   \item{response_linear}{matrix, continuous responses of 2500 people for 100 items generated from linear factor model.}
   \item{response_mirt}{matrix, binary responses of 2500 people for 100 items generated from mirt factor model.}
   \item{response_poi}{matrix, positive integer responses of 2500 people for 100 items generated from Poisson factor model.}}}
\usage{
test_data
}
\description{
A toy example dataset generated from latent factor model.
}
\keyword{datasets}
